test: separate to package
