+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,9 +1,7 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import racingcar.domain.Car
-import racingcar.domain.Distance
 
 class CarTest {
     @Test
